# docs/swagger.yaml
basePath: /api/
definitions:
  apierr.ApiErr:
    properties:
      code:
        type: integer
      message: {}
    type: object
...
  dto.IdResponse:
    properties:
      id:
        example: 00000000-0000-0000-0000-000000000000
        type: string
      message:
        example: readme created successfully
        type: string
    required:
    - id
    type: object
...
  dto.ReadmeResponse:
    properties:
      create_time:
        type: string
      id:
        type: string
      image:
        type: string
      last_update_time:
        type: string
      title:
        type: string
    required:
    - create_time
    - id
    - image
    - last_update_time
    - title
    type: object
...
  dto.SuccessResponse:
    properties:
      code:
        example: 200
        type: integer
      message:
        example: success
        type: string
    type: object
...
  models.Readme:
    properties:
      create_time:
        type: string
      id:
        type: string
      image:
        type: string
      last_update_time:
        type: string
      links:
        items:
          type: string
        type: array
      owner_id:
        type: string
      render_order:
        items:
          type: string
        type: array
      template_id:
        type: string
      text:
        items:
          type: string
        type: array
      title:
        type: string
      widgets:
        items:
          additionalProperties:
            type: string
          type: object
        type: array
    type: object
...
host: localhost:3333
info:
  contact:
    email: readmeow@gmail.com
    name: READMEOW Support
  description: Documentation to READMEOW API
  title: REODMEOW API
  version: 1.0.0
paths:
...
  /api/readmes:
    get:
      description: Returns list of readmes for the authorized user
      parameters:
      - in: query
        minimum: 1
        name: amount
        required: true
        type: integer
      - in: query
        minimum: 1
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            items:
              $ref: '#/definitions/dto.ReadmeResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierr.ApiErr'
      security:
      - ApiKeyAuth: []
      summary: Fetch User Readmes
      tags:
      - Readmes
    post:
      consumes:
      - multipart/form-data
      description: Creating a new readme
      parameters:
      - format: binary
        in: formData
        name: image
        required: true
        type: string
      - collectionFormat: csv
        in: formData
        items:
          type: string
        name: links
        type: array
      - collectionFormat: csv
        in: formData
        items:
          type: string
        minItems: 1
        name: render_order
        required: true
        type: array
      - in: formData
        name: template_id
        type: string
      - collectionFormat: csv
        in: formData
        items:
          type: string
        name: text
        type: array
      - in: formData
        maxLength: 80
        minLength: 1
        name: title
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/dto.IdResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "409":
          description: Already exists
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "422":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierr.ApiErr'
      security:
      - ApiKeyAuth: []
      summary: Create Readme
      tags:
      - Readmes
  /api/readmes/{readme}:
    delete:
      description: Deleting a user readme by its id
      parameters:
      - description: Readme ID
        in: path
        name: readme
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierr.ApiErr'
      security:
      - ApiKeyAuth: []
      summary: Delete Readme
      tags:
      - Readmes
    get:
      description: Returns single readme by id
      parameters:
      - description: Readme ID
        in: path
        name: readme
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/models.Readme'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierr.ApiErr'
      security:
      - ApiKeyAuth: []
      summary: Get Readme by ID
      tags:
      - Readmes
    patch:
      consumes:
      - multipart/form-data
      description: Updating existing readme by id
      parameters:
      - in: formData
        name: id
        required: true
        type: string
      - format: binary
        in: formData
        name: image
        required: true
        type: string
      - collectionFormat: csv
        in: formData
        items:
          type: string
        name: links
        type: array
      - collectionFormat: csv
        in: formData
        items:
          type: string
        minItems: 1
        name: render_order
        required: true
        type: array
      - collectionFormat: csv
        in: formData
        items:
          type: string
        name: text
        type: array
      - in: formData
        maxLength: 80
        minLength: 1
        name: title
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "422":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierr.ApiErr'
      security:
      - ApiKeyAuth: []
      summary: Update Readme
      tags:
      - Readmes
...
swagger: "2.0"